# api/dist/modules/hotel/docs/swagger-hotel.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hotel Service API",
    "description": "Tài liệu API của riêng Module Khách sạn",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/api/hotels/{id}": {
      "get": {
        "tags": [
          "Hotels"
        ],
        "summary": "Lấy chi tiết một khách sạn theo ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Hotels"
        ],
        "summary": "Cập nhật thông tin khách sạn",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHotelDto"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hotels"
        ],
        "summary": "Xóa mềm khách sạn",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateHotelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Khách sạn Mường Thanh"
          },
          "description": {
            "type": "string",
            "example": "Khách sạn 5 sao..."
          },
          "address": {
            "type": "string",
            "example": "123 Đường Trần Phú"
          },
          "city": {
            "type": "string",
            "example": "Nha Trang"
          },
          "country": {
            "type": "string",
            "example": "Vietnam"
          },
          "amenities": {
            "type": "array",
            "example": [
              "WIFI",
              "SWIMMING_POOL"
            ],
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          }
        }
      },
...
    }
  }
}